requirements.txt
crewai[openai]==1.3.0
streamlit>=1.32.0
python-dotenv>=1.1.1
pandas>=2.2.0
requests>=2.31.0
openai>=1.51.0
tenacity>=8.2.0
